n turns

+++ include/grid_funcs.h
@@ -87,7 +87,7 @@ void v3d_create_skyrmion_at_old(v3d *v, unsigned int rows, unsigned int cols, in
 void v3d_create_skyrmion_at(v3d *v, unsigned int rows, unsigned int cols, double radius, double dw_width, double ix, double iy, double Q, double vorticity, double _gamma);
 void v3d_create_biskyrmion_at(v3d *v, unsigned int rows, unsigned int cols, double radius, double dw_width, double ix, double iy, double dr, double angle, double Q, double vorticity, double _gamma);
 void v3d_create_skyrmionium_at(v3d *v, unsigned int rows, unsigned int cols, double radius, double dw_width, double ix, double iy, double Q, double vorticity, double _gamma);
-void v3d_create_target_skyrmion_at(v3d *v, unsigned int rows, unsigned int cols, double radius, double dw_width, double ix, double iy, double Q, double vorticity, double _gamma, int turns);
+void v3d_create_target_skyrmion_at(v3d *v, unsigned int rows, unsigned int cols, double radius, double dw_width, double ix, double iy, double Q, double vorticity, double _gamma, double turns);
 void v3d_uniform(v3d *v, unsigned int rows, unsigned int cols, v3d dir);
 
 void grid_fill_with_random(grid *g);
@@ -95,7 +95,7 @@ void grid_create_skyrmion_at_old(grid *g, int radius, int row, int col, double Q
 void grid_create_skyrmion_at(grid *g, double radius, double dw_width, double ix, double iy, double Q, double vorticity, double _gamma);
 void grid_create_biskyrmion_at(grid *g, double radius, double dw_width, double ix, double iy, double dr, double angle, double Q, double vorticity, double _gamma);
 void grid_create_skyrmionium_at(grid *g, double radius, double dw_width, double ix, double iy, double Q, double vorticity, double _gamma);
-void grid_create_target_skyrmion_at(grid *g, double radius, double dw_width, double ix, double iy, double Q, double vorticity, double _gamma, int turns);
+void grid_create_target_skyrmion_at(grid *g, double radius, double dw_width, double ix, double iy, double Q, double vorticity, double _gamma, double turns);
 void grid_uniform(grid *g, v3d dir);
 
 bool grid_free(grid *g);
